Updates role to backend developer
Refactors the profile to reflect a focus on backend development.

This commit updates the main title and description in both English and Spanish versions of the header and about sections, replacing "Full Stack Developer" with "Backend Developer". Additionally, it updates CSS file paths and adds PDF files containing certifications and recommendations. It introduces a demo card with a demo button to display the Spring task manager.